Move UpdateCSharpBuffer
