+++ src/Razor/src/Microsoft.VisualStudio.LanguageServerClient.Razor/DefaultRazorLanguageServerCustomMessageTarget.cs
@@ -126,58 +126,6 @@ internal partial class DefaultRazorLanguageServerCustomMessageTarget : RazorLang
         _documentSynchronizer = documentSynchronizer;
     }
 
-    public override async Task UpdateCSharpBufferAsync(UpdateBufferRequest request, CancellationToken cancellationToken)
-    {
-        if (request is null)
-        {
-            throw new ArgumentNullException(nameof(request));
-        }
-
-        await _joinableTaskFactory.SwitchToMainThreadAsync(cancellationToken);
-
-        UpdateCSharpBuffer(request);
-    }
-
-    // Internal for testing
-    internal void UpdateCSharpBuffer(UpdateBufferRequest request)
-    {
-        if (request is null || request.HostDocumentFilePath is null || request.HostDocumentVersion is null)
-        {
-            return;
-        }
-
-        var hostDocumentUri = new Uri(request.HostDocumentFilePath);
-
-        if (_documentManager.TryGetDocument(hostDocumentUri, out var documentSnapshot) &&
-            documentSnapshot.TryGetAllVirtualDocuments<CSharpVirtualDocumentSnapshot>(out var virtualDocuments))
-        {
-            foreach (var virtualDocument in virtualDocuments)
-            {
-                // TODO: Remove the null check from this line when multiple CSharpVirtualDocuments are actually being created
-                if (virtualDocument.ProjectKey.Id == request.ProjectKeyId || virtualDocument.ProjectKey.Id is null)
-                {
-                    _documentManager.UpdateVirtualDocument<CSharpVirtualDocument>(
-                        hostDocumentUri,
-                        virtualDocument.Uri,
-                        request.Changes.Select(change => change.ToVisualStudioTextChange()).ToArray(),
-                        request.HostDocumentVersion.Value,
-                        state: null);
-                    return;
-                }
-            }
-
-            // If the particular document supports multiple virtual documents, we don't want to try to update a single one
-            Debug.Fail("Multiple virtual documents seem to be supported, but none were updated.");
-            return;
-        }
-
-        _documentManager.UpdateVirtualDocument<CSharpVirtualDocument>(
-            hostDocumentUri,
-            request.Changes.Select(change => change.ToVisualStudioTextChange()).ToArray(),
-            request.HostDocumentVersion.Value,
-            state: null);
-    }
-
     public override async Task UpdateHtmlBufferAsync(UpdateBufferRequest request, CancellationToken cancellationToken)
     {
         if (request is null)
